fix(decodingSignature): Fix to investiguate on the issue #36
The regexp was too eager, so it lead to a syntax error. I've changed the regexp to get only the line involded in decoding part.
Due to the lack of unit-testing in the project, I can't assure this doesn't break other "download"

Linked to #36 and testing against video submitted by @ahmadftx and @TigranTsat and of course @davinkevin